refactor: extract presentation feature folder
Move presentation engine, presentation-decorations, screen-wake-lock,
slideshow + teleprompter widgets, and slideshow + teleprompter screens
into src/app/features/presentation/. Expose presentationExtensions
factory for the editor extension slot.
