+++ src/app/features/presentation/index.ts
@@ -1,0 +1,33 @@
+export {
+	parsePresentation,
+	getPresentationMode,
+	parsePresentationSize,
+	parsePresentationTheme,
+} from "./lib/presentation"
+export type {
+	PresentationItem,
+	VisualBlock,
+	SlideContent,
+	PresentationSize,
+	PresentationTheme,
+	TextSegment,
+} from "./lib/presentation"
+export { useScreenWakeLock } from "./lib/screen-wake-lock"
+export { presentationExtensions } from "./lib/editor-extension"
+export { Slideshow } from "./widgets/slideshow"
+export type { Slide, HighlightRange } from "./widgets/slideshow"
+export { Teleprompter, groupBySlide } from "./widgets/teleprompter"
+export type { SlideGroup } from "./widgets/teleprompter"
+export { SidebarPresentationLinks } from "./widgets/sidebar-presentation-links"
+export {
+	SlideshowScreen,
+	resolve as slideshowResolve,
+	loadWikilinkCache as loadSlideshowWikilinkCache,
+} from "./screens/slideshow-screen"
+export type { LoaderData as SlideshowLoaderData } from "./screens/slideshow-screen"
+export {
+	TeleprompterScreen,
+	resolve as teleprompterResolve,
+	loadWikilinkCache as loadTeleprompterWikilinkCache,
+} from "./screens/teleprompter-screen"
+export type { LoaderData as TeleprompterLoaderData } from "./screens/teleprompter-screen"
