+++ src/app/features/presentation/lib/editor-extension.ts
@@ -1,0 +1,8 @@
+import { type Extension } from "@codemirror/state"
+import { createPresentationDecorations } from "./presentation-decorations"
+
+export { presentationExtensions }
+
+function presentationExtensions(): Extension[] {
+	return [createPresentationDecorations()]
+}
