+++ src/app/routes/doc.$id.index.tsx
@@ -70,6 +70,10 @@ import { SidebarProvider, useSidebar } from "@/components/ui/sidebar"
 import { HelpCircle, Loader2, Search, Settings, Plus } from "lucide-react"
 
 import { SidebarViewLinks } from "@/components/sidebar-view-links"
+import {
+	SidebarPresentationLinks,
+	presentationExtensions,
+} from "@/app/features/presentation"
 import { SidebarFileMenu } from "@/components/sidebar-file-menu"
 import { SidebarEditMenu } from "@/components/sidebar-edit-menu"
 import { SidebarFormatMenu } from "@/components/sidebar-format-menu"
@@ -512,6 +516,7 @@ function EditorContent({ doc, docId }: EditorContentProps) {
 							getAssets: () =>
 								assetsRef.current.map(a => ({ id: a.id, name: a.name })),
 						}),
+						...presentationExtensions(),
 					]}
 				/>
 				<EditorToolbar
@@ -569,6 +574,7 @@ function EditorContent({ doc, docId }: EditorContentProps) {
 							</SidebarMenuItem>
 							<SidebarSeparator />
 							<SidebarViewLinks doc={doc} />
+							<SidebarPresentationLinks doc={doc} />
 							<SidebarSeparator />
 							<SidebarFileMenu
 								doc={doc}
