+++ src/lib/doc-resolver.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react"
 import { Document } from "@/schema"
 import { getDocumentTitle } from "@/lib/document-utils"
-import { getPresentationMode } from "@/lib/presentation"
+import { getPresentationMode } from "@/app/features/presentation"
 
 export { resolveDocTitle, resolveDocTitles, useDocTitle, useDocTitles }
 export type { ResolvedDoc }
